+++ packages/cache/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @prairielearn/cache
 
+## 2.1.0
+
+### Minor Changes
+
+- e1c7f75: Add support for an optional type parameter to `cache.get(...)`
+
+## 2.0.24
+
+### Patch Changes
+
+- 2f5ce1f: Upgrade all JavaScript dependencies
+- Updated dependencies [2f5ce1f]
+  - @prairielearn/sentry@4.0.4
+
 ## 2.0.23
 
 ### Patch Changes